feat: open Custom OAuth in a native browser session to support WebAuthn/passkeys
The in-app WebView (AuthenticationWebView) cannot run the WebAuthn API, so
passkey / security-key login through a Custom OAuth provider (e.g. Keycloak)
fails. Open Custom OAuth in a native browser session instead —
ASWebAuthenticationSession on iOS, Chrome Custom Tabs on Android — via
expo-web-browser's openAuthSessionAsync, which runs in a real browser context
where WebAuthn is available.

onPressCustomOAuth now uses the 'redirect' login style (so the server redirects
to the existing rocketchat://auth callback) and openAuthSessionAsync, which
resolves directly with the redirect URL. parseDeepLinking is extracted from
app/index.tsx into a shared helper and now tags
rocketchat://auth?credentialToken=... redirects as type 'oauth', activating the
existing-but-dormant handleOAuth deep-linking saga to complete the login.

No new dependency (expo-web-browser is already used by openLink), no native
config changes (the rocketchat:// URL scheme is already registered on iOS and
Android), and the deep-linking saga is unchanged — the native session keeps the
app foregrounded and connected, so no reconnection handling is needed.

Closes #5681
