+++ app/index.tsx
@@ -22,7 +22,7 @@ import { MIN_WIDTH_MASTER_DETAIL_LAYOUT } from './lib/constants/tablet';
 import { getAllowAnalyticsEvents, getAllowCrashReport } from './lib/methods/crashReport';
 import { debounce, isTablet } from './lib/methods/helpers';
 import { toggleAnalyticsEventsReport, toggleCrashErrorsReport } from './lib/methods/helpers/log';
-import parseQuery from './lib/methods/helpers/parseQuery';
+import parseDeepLinking from './lib/methods/helpers/parseDeepLinking';
 import {
 	getTheme,
 	initialTheme,
@@ -63,29 +63,6 @@ interface IState {
 	scale: number;
 	fontScale: number;
 }
-
-const parseDeepLinking = (url: string) => {
-	if (url) {
-		url = url.replace(/rocketchat:\/\/|https:\/\/go.rocket.chat\//, '');
-		const regex = /^(room|auth|invite|shareextension)\?/;
-		const match = url.match(regex);
-		if (match) {
-			const matchedPattern = match[1];
-			const query = url.replace(regex, '').trim();
-
-			if (query) {
-				const parsedQuery = parseQuery(query);
-				return {
-					...parsedQuery,
-					type: matchedPattern === 'shareextension' ? matchedPattern : parsedQuery?.type
-				};
-			}
-		}
-	}
-
-	// Return null if the URL doesn't match or is not valid
-	return null;
-};
 
 export default class Root extends React.Component<{}, IState> {
 	private listenerTimeout!: any;
